fix(agent): resolve Unknown connection_id — UUID string comparison, dual-key connections HashMap, settings.connections sent
Three bugs fixed in the agent connection resolution chain:

1. **UUID vs number type mismatch**: DatabaseSource.connectionId was typed
   as number but connection IDs are UUID strings. Number(uuid) = NaN, so
   contextProvider() could never match an attached source to its connection.
   Fixed by changing connectionId from number to string everywhere.

2. **connections HashMap keyed by alias, LLM uses UUID**: The connectionConfig
   was keyed by alias ("mac-postgresql") but the LLM passes the UUID
   ("5311309e-...") from sqlkit__list_connections output as connection_id.
   Fixed by keying entries by BOTH UUID and alias via flatMap.

3. **settings.connections never sent to backend**: The loop runner reads
   settings["connections"] to build its HashMap, but only connectionConfig
   (camelCase) was sent. Added settings.connections = connectionConfig.
