--- src-tauri/src/agent/config.rs
@@ -1,1 +1,0 @@
-pub use crate::agent::provider_adapter::{build_headers, get_base_url};
