+++ src-tauri/src/agent/mod.rs
@@ -1,18 +1,4 @@
-pub mod chat_formatter;
-pub mod compact;
-pub mod config;
-pub mod conversation;
+// App-specific agent modules — shared logic is in data-studio-agent-lib.
 pub mod executor;
-pub mod harness;
-pub mod loop_runner;
-pub mod loop_runner_support;
-pub mod model_registry;
-pub mod provider_adapter;
 pub mod query_history;
 pub mod session_store;
-pub mod token_counter;
-pub mod tool_executor;
-pub mod tools;
-
-pub use harness::{list_llm_models, run_agent_step, validate_llm_config};
-pub use tools::get_all_tools;
