+++ src-tauri/src/agent/query_history.rs
@@ -1,7 +1,7 @@
 use serde::{Deserialize, Serialize};
 use tauri::State;
 
-use crate::db::AgentDb;
+use data_studio_agent_storage_sqlite::db::AgentDb;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct QueryHistoryEntry {
